Basic heading implementation
This commit makes adjustments to the Odometry class. The class is now able to output heading values that roughly in the correct ballpark and also provides independent readings for wheel velocities. Some work into calibration is needed as the error from the wheels are ~ +/- 5 degrees from expected values. Further testing is definitely needed to better quantify the error